move the "with-trpc" example over to file-based routing
