+++ examples/react/with-trpc/app/routeTree.gen.ts
@@ -11,17 +11,105 @@
 // Import Routes
 
 import { Route as rootRoute } from './routes/__root'
+import { Route as DashboardImport } from './routes/dashboard'
+import { Route as IndexImport } from './routes/index'
+import { Route as DashboardIndexImport } from './routes/dashboard.index'
+import { Route as DashboardPostsImport } from './routes/dashboard.posts'
+import { Route as DashboardPostsIndexImport } from './routes/dashboard.posts.index'
+import { Route as DashboardPostsPostIdImport } from './routes/dashboard.posts.$postId'
 
 // Create/Update Routes
+
+const DashboardRoute = DashboardImport.update({
+  path: '/dashboard',
+  getParentRoute: () => rootRoute,
+} as any)
+
+const IndexRoute = IndexImport.update({
+  path: '/',
+  getParentRoute: () => rootRoute,
+} as any)
+
+const DashboardIndexRoute = DashboardIndexImport.update({
+  path: '/',
+  getParentRoute: () => DashboardRoute,
+} as any)
+
+const DashboardPostsRoute = DashboardPostsImport.update({
+  path: '/posts',
+  getParentRoute: () => DashboardRoute,
+} as any)
+
+const DashboardPostsIndexRoute = DashboardPostsIndexImport.update({
+  path: '/',
+  getParentRoute: () => DashboardPostsRoute,
+} as any)
+
+const DashboardPostsPostIdRoute = DashboardPostsPostIdImport.update({
+  path: '/$postId',
+  getParentRoute: () => DashboardPostsRoute,
+} as any)
 
 // Populate the FileRoutesByPath interface
 
 declare module '@tanstack/react-router' {
-  interface FileRoutesByPath {}
+  interface FileRoutesByPath {
+    '/': {
+      id: '/'
+      path: '/'
+      fullPath: '/'
+      preLoaderRoute: typeof IndexImport
+      parentRoute: typeof rootRoute
+    }
+    '/dashboard': {
+      id: '/dashboard'
+      path: '/dashboard'
+      fullPath: '/dashboard'
+      preLoaderRoute: typeof DashboardImport
+      parentRoute: typeof rootRoute
+    }
+    '/dashboard/posts': {
+      id: '/dashboard/posts'
+      path: '/posts'
+      fullPath: '/dashboard/posts'
+      preLoaderRoute: typeof DashboardPostsImport
+      parentRoute: typeof DashboardImport
+    }
+    '/dashboard/': {
+      id: '/dashboard/'
+      path: '/'
+      fullPath: '/dashboard/'
+      preLoaderRoute: typeof DashboardIndexImport
+      parentRoute: typeof DashboardImport
+    }
+    '/dashboard/posts/$postId': {
+      id: '/dashboard/posts/$postId'
+      path: '/$postId'
+      fullPath: '/dashboard/posts/$postId'
+      preLoaderRoute: typeof DashboardPostsPostIdImport
+      parentRoute: typeof DashboardPostsImport
+    }
+    '/dashboard/posts/': {
+      id: '/dashboard/posts/'
+      path: '/'
+      fullPath: '/dashboard/posts/'
+      preLoaderRoute: typeof DashboardPostsIndexImport
+      parentRoute: typeof DashboardPostsImport
+    }
+  }
 }
 
 // Create and export the route tree
 
-export const routeTree = rootRoute.addChildren({})
+export const routeTree = rootRoute.addChildren({
+  IndexRoute,
+  DashboardRoute: DashboardRoute.addChildren({
+    DashboardPostsRoute: DashboardPostsRoute.addChildren({
+      DashboardPostsPostIdRoute,
+      DashboardPostsIndexRoute,
+    }),
+    DashboardIndexRoute,
+  }),
+})
 
 /* prettier-ignore-end */
