+++ examples/react/with-trpc/app/routes/-components/spinner.tsx
@@ -1,0 +1,5 @@
+import * as React from 'react'
+
+export function Spinner() {
+  return <div className="inline-block animate-spin px-3">⍥</div>
+}
