+++ examples/react/with-trpc/app/routes/dashboard.tsx
@@ -1,0 +1,48 @@
+import * as React from 'react'
+import { Link, Outlet, createFileRoute } from '@tanstack/react-router'
+
+export const Route = createFileRoute('/dashboard')({
+  component: DashboardComponent,
+})
+
+function DashboardComponent() {
+  return (
+    <>
+      <div className="flex items-center border-b">
+        <h2 className="text-xl p-2">Dashboard</h2>
+        <Link
+          to="/dashboard/posts/$postId"
+          params={{
+            postId: '3',
+          }}
+          className="py-1 px-2 text-xs bg-blue-500 text-white rounded-full"
+        >
+          1 New Invoice
+        </Link>
+      </div>
+      <div className="flex flex-wrap divide-x">
+        {(
+          [
+            ['.', 'Summary'],
+            ['/dashboard/posts', 'Posts'],
+          ] as const
+        ).map(([to, label]) => {
+          return (
+            <Link
+              from={Route.fullPath}
+              key={to}
+              to={to}
+              activeOptions={{ exact: to === '.' }}
+              activeProps={{ className: `font-bold` }}
+              className="p-2"
+            >
+              {label}
+            </Link>
+          )
+        })}
+      </div>
+      <hr />
+      <Outlet />
+    </>
+  )
+}
